+++ packages/common/src/authorization/projectMemberAbility.mock.ts
@@ -1,7 +1,5 @@
-import {
-    ProjectMemberProfile,
-    ProjectMemberRole,
-} from '../types/projectMemberProfile';
+import { ProjectMemberProfile } from '../types/projectMemberProfile';
+import { ProjectMemberRole } from '../types/projectMemberRole';
 
 export const PROJECT_VIEWER: ProjectMemberProfile = {
     userUuid: 'user-uuid-1234',