+++ packages/frontend/src/components/common/modal/DashboardUpdateModal.tsx
@@ -5,6 +5,7 @@ import {
     Modal,
     ModalProps,
     Stack,
+    Textarea,
     TextInput,
     Title,
 } from '@mantine/core';
@@ -75,10 +76,12 @@ const DashboardUpdateModal: FC<DashboardUpdateModalProps> = ({
                         {...form.getInputProps('name')}
                     />
 
-                    <TextInput
+                    <Textarea
                         label="Description"
                         placeholder="A few words to give your team some context"
                         disabled={isUpdating}
+                        autosize
+                        maxRows={3}
                         {...form.getInputProps('description')}
                     />